# chore: Webpack, babel-loader, Prettier, TSLint