--- a/.babelrc
+++ b/.babelrc
@@ -1,0 +1,6 @@
+{
+    "presets": [
+        "@babel/env",
+        "@babel/typescript"
+    ]
+}