--- a/.babelrc
+++ b/.babelrc
@@ -2,5 +2,8 @@
     "presets": [
         "@babel/env",
         "@babel/typescript"
+    ],
+    "plugins": [
+        "@babel/proposal-class-properties"
     ]
 }