--- a/package.json
+++ b/package.json
@@ -56,7 +56,9 @@
     "regenerator-runtime": "^0.13.1",
     "semantic-release": "^15.13.3",
     "typescript": "3.2.4",
-    "web3-core-promievent": "^1.0.0-beta.37"
+    "web3-core-promievent": "^1.0.0-beta.37",
+    "webpack": "^4.29.6",
+    "webpack-cli": "^3.3.0"
   },
   "husky": {
     "hooks": {