--- a/package.json
+++ b/package.json
@@ -6,7 +6,9 @@
   "scripts": {
     "test": "jest",
     "semantic-release": "semantic-release",
-    "postinstall": "rm -f node_modules/web3/index.d.ts node_modules/web3/types.d.ts"
+    "postinstall": "rm -f node_modules/web3/index.d.ts node_modules/web3/types.d.ts",
+    "build:dev": "webpack --mode development",
+    "build:prod": "webpack --mode production"
   },
   "jest": {
     "moduleFileExtensions": [