--- a/package.json
+++ b/package.json
@@ -42,6 +42,7 @@
   },
   "devDependencies": {
     "@babel/core": "^7.2.0",
+    "@babel/preset-env": "^7.4.2",
     "@babel/preset-typescript": "^7.3.3",
     "@commitlint/cli": "^7.5.2",
     "@commitlint/config-conventional": "^7.5.0",
@@ -53,6 +54,7 @@
     "@types/web3": "^1.0.18",
     "babel-core": "^7.0.0-bridge.0",
     "babel-jest": "^23.6.0",
+    "babel-loader": "^8.0.5",
     "husky": "^1.3.1",
     "jest": "^24.1.0",
     "reflect-metadata": "^0.1.12",