--- a/package.json
+++ b/package.json
@@ -7,7 +7,7 @@
     "test": "jest",
     "semantic-release": "semantic-release",
     "postinstall": "rm -f node_modules/web3/index.d.ts node_modules/web3/types.d.ts",
-    "build:dev": "webpack --mode development",
+    "build:dev": "webpack --mode development --display-error-details",
     "build:prod": "webpack --mode production"
   },
   "jest": {