--- a/package.json
+++ b/package.json
@@ -60,6 +60,7 @@
     "reflect-metadata": "^0.1.12",
     "regenerator-runtime": "^0.13.1",
     "semantic-release": "^15.13.3",
+    "tsconfig-paths-webpack-plugin": "^3.2.0",
     "typescript": "3.2.4",
     "web3-core-promievent": "^1.0.0-beta.37",
     "webpack": "^4.29.6",