--- a/package.json
+++ b/package.json
@@ -58,6 +58,7 @@
     "babel-loader": "^8.0.5",
     "husky": "^1.3.1",
     "jest": "^24.1.0",
+    "lint-staged": "^8.1.5",
     "prettier": "1.16.4",
     "reflect-metadata": "^0.1.12",
     "regenerator-runtime": "^0.13.1",
@@ -70,8 +71,15 @@
   },
   "husky": {
     "hooks": {
-      "commit-msg": "commitlint -E HUSKY_GIT_PARAMS"
+      "commit-msg": "commitlint -E HUSKY_GIT_PARAMS",
+      "pre-commit": "lint-staged"
     }
+  },
+  "lint-staged": {
+    "*.{js,json,css,md}": [
+      "prettier --write",
+      "git add"
+    ]
   },
   "repository": {
     "type": "git",