--- a/package.json
+++ b/package.json
@@ -65,6 +65,7 @@
     "semantic-release": "^15.13.3",
     "tsconfig-paths-webpack-plugin": "^3.2.0",
     "tslint": "^5.15.0",
+    "tslint-config-prettier": "^1.18.0",
     "typescript": "3.2.4",
     "web3-core-promievent": "^1.0.0-beta.37",
     "webpack": "^4.29.6",