--- a/package.json
+++ b/package.json
@@ -78,7 +78,7 @@
     }
   },
   "lint-staged": {
-    "*.{js,json,css,md}": [
+    "*.{ts,tsx,js,jsx,json,css,md}": [
       "prettier --write",
       "git add"
     ]