--- a/tslint.json
+++ b/tslint.json
@@ -1,0 +1,3 @@
+{
+  "extends": ["tslint:latest", "tslint-config-prettier"]
+}