--- a/webpack.config.js
+++ b/webpack.config.js
@@ -1,13 +1,19 @@
+const { TsconfigPathsPlugin } = require('tsconfig-paths-webpack-plugin');
+
 module.exports = {
     entry: './src/index.ts',
+    resolve: {
+        extensions: ['.tsx', '.ts', '.js'],
+        plugins: [
+            new TsconfigPathsPlugin()
+        ]
+    },
     module: {
         rules: [
             {
                 test: /\.tsx?$/,
                 exclude: /node_modules/,
-                use: {
-                    loader: "babel-loader"
-                }
+                loader: "babel-loader"
             }
         ]
     }