@fernandoespinosa 2019-04-05
:tada: This PR is included in version 1.0.0 :tada:

The release is available on:
- [npm package (@latest dist-tag)](https://www.npmjs.com/package/@polymathnetwork/sdk)
- [GitHub release](https://github.com/PolymathNetwork/polymath-sdk/releases/tag/v1.0.0)

Your **[semantic-release](https://github.com/semantic-release/semantic-release)** bot :package::rocket: